REPLAY-1556 Do not mask tab selection for MASK_ALL privacy
